changes size_t to int32_t in EMsoftCEBSDDI wrapper header file; part of the solution to issue #25.

+++ Source/EMsoftWrapperLib/DictionaryIndexing/EMsoftDIwrappers.h
@@ -84,8 +84,8 @@ void EMsoftCEBSDDIpreview
 */
 
 void EMsoftCEBSDDI
-	(size_t* ipar, float* fpar, char* spar, float* dpatterns, float* epatterns, 
-	 float* resultmain, size_t* indexmain, ProgCallBackTypeDI3 callback, ProgCallBackTypeError errorcallback,
+	(int32_t* ipar, float* fpar, char* spar, float* dpatterns, float* epatterns, 
+	 float* resultmain, int32_t* indexmain, ProgCallBackTypeDI3 callback, ProgCallBackTypeError errorcallback,
      size_t object, bool* cancel); 
 
 /**
